many updates
Improved logfile handling. The loading of logfile was taking too long to do and I needed to change how the logfiles were saved to be able to use pandas to prse the data. Pandas is writtin in C so it's execution is going to be a heck of a lot faster than parsing the file in Python code. It also handles reading timestamps. This works out far better. I also change how the log files were being displayed. Instead of using a text editor type of control I opted to use a list type of control where I was able to load only the data that is actually viewable. This allows for everything to run a lot smoother when scrolling through the data. Much faster than it was originally and the new one allows you to drill down to a specific hour.

The slash screen also had some issues I have fixes.The text was not properly being displayed and there was also flickering. When initially loading the databases where a lot of updates were being made to the splash screen this caused it to take a long time to load the data into the database. What I changed was i went with using a memory DC to render to a bitmap. The graphics portion which is static is now only rendered 1 time. The area where the test is rendered to now only updates by rendering a black box to cover the old text and then rendering the text after that. All the paint event has to do is render the bitmap to the screen. This eliminates the flickering that was occurring and also speeds things up. I just now thought about this and I can probably make it run even faster if I render the graphics only a single time to the client and then only render a small bitmap for the text area. I might poke about it again to get the performance even better.

I separated the database data from the main program. I did this because the size of the data was getting pretty large. I also changed the code that handles the images, datasheets cads and models so it will collect those things the first time a part gets added to be used instead of having it all loaded beforehand. Loading it all prior to use would have made the application very large in size. It's already sitting at 1.6gb in size and I didn't want to make it any larger. The data that is to be pre loaded is in a separate reporitory.

I fixes the axis indicators so they have been added back in.

+++ harness_designer/database/db_connectors/db_data.py
@@ -22,12 +22,9 @@ class DBData:
             cur_size = 0
 
             splash.SetText(f'Downloading Database Data {cur_size}')
+            splash.flush()
 
             tmpdir = tempfile.gettempdir()
-            tmp_zipfile = os.path.join(tmpdir, 'harness_designer_data.zip')
-
-            if os.path.exists(tmp_zipfile):
-                os.remove(tmp_zipfile)
 
             tmpzipdata = os.path.join(tmpdir, 'harness_designer.database_data')
 
@@ -36,23 +33,29 @@ class DBData:
 
             os.makedirs(tmpzipdata)
 
-            with open(tmp_zipfile, "wb") as file:
-                for data in response.iter_content(block_size):
-                    if not data:
-                        break
+            tmp_zipfile = os.path.join(tmpdir, 'harness_designer_data.zip')
 
-                    cur_size += len(data)
-                    file.write(data)
-                    splash.SetText(f'Downloading Database Data {cur_size}')
+            if not os.path.exists(tmp_zipfile):
+
+                with open(tmp_zipfile, "wb") as file:
+                    for data in response.iter_content(block_size):
+                        if not data:
+                            break
+
+                        cur_size += len(data)
+                        file.write(data)
+                        splash.SetText(f'Downloading Database Data {cur_size}')
 
             splash.SetText(f'Decompressing Database Data...')
+            splash.flush()
 
             with zipfile.ZipFile(tmp_zipfile) as zdata:
                 zdata.extractall(tmpzipdata)
 
             splash.SetText('Finished Decompressing Data!')
+            splash.flush()
 
-            os.remove(tmp_zipfile)
+            # os.remove(tmp_zipfile)
             self._data_dir = tmpzipdata
 
         return os.path.join(self._data_dir, 'database-main')
